Use Default Initialization for Structs
There's no need to have a constructor called for a struct,
we can use the default initialization, that's faster as there's
no function call, less code to type and easier to read.

Considering that the constructor in this case only setted
the internal values in the same order as they where defined,
there's no loss changing to a default initialization, only gains.