Adapt gstreamer integration and plugins for VideoSurface
Here I need to change a few names internally on gstreamerintegration
as the VideoSurface will take care of some things. The pipeline creation
how has a common tail, the Qml manages the Video and we generally test
more things before blindly using them.

The plugins now return a smaller GStreamer pipeline, so we can also
plug tee's in it later.

I also register the new Qml components, and remove the usage of the
dreadful Window component.