Add a AbstractPluginModel
This allow us to extend the plugin model with different
plugins in the future - I plan to start a Plugin for
interface elements.

This code decouples anything that's plugin-irrelevant
from the VehicleSupportPluginModel, loading, unloading, etc.

The VehicleSupportPluginModel now only deals with data()
and with creating the plugin *if* the plugin is VehicleSupportPlugin